Add extra safegaurd for deleting guest accounts

+++ client/src/features/user/routes/User.tsx
@@ -8,6 +8,7 @@ import { notify } from "@/utils/notify";
 import { supabase } from "@/features/auth/lib/auth";
 import { useNavigate } from "react-router-dom";
 import { deleteAccount } from "../api";
+import { getUid } from "@/features/calendar/utils/getUid";
 
 export const User = () => {
   const navigate = useNavigate();
@@ -20,7 +21,7 @@ export const User = () => {
   const signOut = async (e: React.SyntheticEvent) => {
     e.preventDefault();
 
-    if (sessionStorage.getItem("uid")) {
+    if (sessionStorage.getItem("uid") && !(await getUid())) {
       deleteAccount();
       navigate("/account/login");
     }
